Rename package project and namespace
