+++ src/EmotionalCities.Pupil/FrameToSingle.cs
@@ -4,17 +4,17 @@ using System.Linq;
 using System.Reactive.Linq;
 using NetMQ;
 
-namespace PupilInterface
+namespace EmotionalCities.Pupil
 {
-    public class FrameToUInt64 : Transform<NetMQFrame, ulong>
+    public class FrameToSingle : Transform<NetMQFrame, float>
     {
         public int Offset { get; set; } = 0;
 
-        public override IObservable<ulong> Process(IObservable<NetMQFrame> source)
+        public override IObservable<float> Process(IObservable<NetMQFrame> source)
         {
             return source.Select(frame =>
             {
-                return BitConverter.ToUInt64(frame.Buffer, Offset);
+                return BitConverter.ToSingle(frame.Buffer, Offset);
             });
         }
     }
