+++ src/EmotionalCities.Pupil/Properties/launchSettings.json
